// src/app/models/juegos.ts
export class juegos{

    id: number;
    slug: string;
    name: string;
    name_original: string;
    description: string;
    released: string;
    tba: boolean;
    background_image: string;
    rating: number;
    platforms: [];
    genres: [];
    developers: [];
    publishers: [];
    clip: string;
    preview: any;

}