// src/app/models/registro.ts
export class registro{
    
    // constructor(name: string,email: string,lugar: string,cedula: number){
    //     this.name= name
    //     this.email= email
    //     this.lugar= lugar
    //     this.cedula= cedula
    // }
    name: string;
    email: string;
    lugar: string;
    cedula: number;
 
}